some progress

+++ c/temperature_conversion.h
@@ -2,7 +2,7 @@
 #ifndef TEMPERATURE_CONVERSION_H
 #define TEMPERETURE_CONVERSION_H
 
-int NTC_to_celsius(unsigned char ntc);
+float NTC_to_celsius(unsigned char ntc);
 
 unsigned char celsius_to_NTC(int celsius);
